+++ Changelog.md
@@ -7,6 +7,7 @@
 ### Australia refactor/fixes
 
 * Australian Capital Territory fixes: added the Easter Sunday to holidays.
+* New South Wales: Fixed the ANZAC day shift flag. This is not a holiday if it falls on a week-end.
 
 ## v15.3.0 (2021-05-07)